+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **[Breaking Change]** `emit_swf` now returns `Result<(), SwfEmitError>`.
 - **[Change]** Unsupported compression methods no longer panic, they return an error instead.
 - **[Feature]** Add support for DEFLATE compression with the `miniz_oxide` crate.
+- **[Feature]** Add support for LZMA compression with the `lzma-rs` crate.
 
 # 0.13.0 (2021-07-24)